updating security authentication
