+++ src/main/java/ultigreen/demo/controller/APIUserAuthentication.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import ultigreen.demo.domain.User;
+import ultigreen.demo.domain.AppUser;
 import ultigreen.demo.domain.UserRepository;
 
 @RestController
@@ -23,7 +23,7 @@ public class APIUserAuthentication {
     	}
     
     @PostMapping("/sign-up")
-    public void signUp(@RequestBody User user) {
+    public void signUp(@RequestBody AppUser user) {
         user.setPassword(bCryptPasswordEncoder.encode(user.getPassword()));
         userRepository.save(user);
     }
