+++ src/main/java/ultigreen/demo/domain/UserRepository.java
@@ -2,8 +2,8 @@ package ultigreen.demo.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Long>{
+public interface UserRepository extends JpaRepository<AppUser, Long>{
 
-	User findByUsername(String username);
+	AppUser findByUsername(String username);
 
 }
